flask_docker/requirements.txt
Flask==1.1.1
Gunicorn
Flask-Cors==3.0.8
kafka-python